refactor(tests): split company.test.ts into focused files under 400 lines
